refactor(vision): split monolithic base.py into three base files
Split the former vision/base.py into three files by responsibility:

- vision/base.py         — data transfer protocol only
                          (SpatialLayout, PhysicalMask, ProcessorOutput)
- vision/segmenters/base.py — Segmenter(ABC), SegmenterConfig,
                          per-strategy param dataclasses (PatchParams,
                          SlicParams, GradientGuidedParams, CustomSegmenterParams)
- vision/maskers/base.py    — Masker(ABC), MaskerConfig,
                          per-strategy param dataclasses (CrossModalMeanParams,
                          CrossModalBlurParams, VisionMeanParams,
                          VisionBlurParams, TextAttentionParams)

All cross-references within the vision package updated to relative imports.
99 tests pass.
