+++ src/shapiq/imputer/vision/base.py
@@ -1,145 +1,21 @@
-"""Base module for the vision imputer sub-package.
+"""
+Base module for the vision imputer sub-package — data transfer protocol.
 
-Defines:
-    - Data transfer types: SegmenterConfig, MaskerConfig, SpatialLayout,
-      PhysicalMask, ProcessorOutput, and per-strategy params.
-    - Abstract contracts: Segmenter(ABC), Masker(ABC).
+Defines data types shared by all sub-modules:
+    - SpatialLayout — describes how input is divided into players
+    - PhysicalMask — concrete pixel/token-level masks
+    - ProcessorOutput — standardised HuggingFace processor wrapper
+
+Segmenter and Masker abstract contracts now live in:
+    - shapiq.imputer.vision.segmenters.base
+    - shapiq.imputer.vision.maskers.base
 """
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
-import numpy as np
 import torch
-
-# ═══════════════════════════════════════════════════════════════════════
-# Segmenter Configuration
-# ═══════════════════════════════════════════════════════════════════════
-
-
-@dataclass
-class PatchParams:
-    """Rigid-grid patch segmenter parameters.  No configurable knobs."""
-
-
-@dataclass
-class SlicParams:
-    """SLIC superpixel segmentation parameters.
-
-    Attributes:
-        n_segments: Target superpixel count (default 49, ≈ 7×7 grid).
-        compactness: SLIC compactness factor (higher = more regular shapes).
-        sigma: Pre-segmentation Gaussian blur sigma.
-    """
-
-    n_segments: int = 49
-    compactness: float = 10.0
-    sigma: float = 0.0
-
-
-@dataclass
-class GradientGuidedParams:
-    """Gradient-guided saliency segmentation parameters.
-
-    Attributes:
-        n_segments: Target superpixel count. None means derive from
-            grid_size (ViT) or fall back to 49.
-    """
-
-    n_segments: int | None = None
-
-
-@dataclass
-class CustomSegmenterParams:
-    """User-provided binary mask segmenter parameters.
-
-    Each mask defines one player. Masks are supplied directly to
-    ``CustomSegmenter`` at construction time, not via this dataclass.
-    """
-
-
-@dataclass
-class SegmenterConfig:
-    """Complete configuration for a Segmenter.
-
-    Caller-provided: strategy + per-strategy params (patch / slic / gradient_guided).
-    Factory-populated: model metadata (image_size, patch_size, model_type, ...).
-    Default strategy is ``"patch"``.
-    """
-
-    strategy: str = "patch"
-    patch: PatchParams = field(default_factory=PatchParams)
-    slic: SlicParams = field(default_factory=SlicParams)
-    gradient_guided: GradientGuidedParams = field(default_factory=GradientGuidedParams)
-    custom_segmenter: CustomSegmenterParams = field(default_factory=CustomSegmenterParams)
-
-    # Factory-populated (model metadata)
-    model_type: str = ""
-    image_size: int = 0
-    patch_size: int = 0
-    n_channels: int = 3
-    grid_size: int = 0
-    n_players_image: int = 0
-    n_players_text: int = 0
-    text_total_length: int = 0
-
-    @property
-    def active_params(self):
-        return getattr(self, self.strategy, None)
-
-
-# ═══════════════════════════════════════════════════════════════════════
-# Masker Configuration
-# ═══════════════════════════════════════════════════════════════════════
-
-
-@dataclass
-class CrossModalMeanParams:
-    """Cross-modal occlusion (vision-mean + text-attention)."""
-
-
-@dataclass
-class CrossModalBlurParams:
-    """Cross-modal occlusion (vision-blur + text-attention)."""
-
-
-@dataclass
-class VisionMeanParams:
-    """Pure image occlusion via multiplicative binary mask."""
-
-
-@dataclass
-class VisionBlurParams:
-    """Pure image occlusion via Gaussian blur."""
-
-    sigma: float = 3.0
-
-
-@dataclass
-class TextAttentionParams:
-    """Pure text occlusion via attention_mask replacement."""
-
-
-@dataclass
-class MaskerConfig:
-    """Complete configuration for a Masker.
-
-    Caller-provided: strategy + per-strategy params.
-    Default strategy is ``"crossmodal_mean"``.
-    """
-
-    strategy: str = "crossmodal_mean"
-    crossmodal_mean: CrossModalMeanParams = field(default_factory=CrossModalMeanParams)
-    crossmodal_blur: CrossModalBlurParams = field(default_factory=CrossModalBlurParams)
-    vision_mean: VisionMeanParams = field(default_factory=VisionMeanParams)
-    vision_blur: VisionBlurParams = field(default_factory=VisionBlurParams)
-    text_attn: TextAttentionParams = field(default_factory=TextAttentionParams)
-
-    @property
-    def active_params(self):
-        return getattr(self, self.strategy, None)
 
 
 # ═══════════════════════════════════════════════════════════════════════
@@ -252,121 +128,3 @@ class ProcessorOutput:
         self.input_ids = self.input_ids.to(device)
         self.attention_mask = self.attention_mask.to(device)
         return self
-
-
-# ═══════════════════════════════════════════════════════════════════════
-# Abstract Contracts
-# ═══════════════════════════════════════════════════════════════════════
-
-
-class Segmenter(ABC):
-    """Abstract base class for spatial division strategies.
-
-    A Segmenter defines which pixels/tokens belong to which player.
-    It produces a SpatialLayout and converts coalition arrays into
-    PhysicalMask tensors.
-    """
-
-    def __init__(self, config: SegmenterConfig):
-        self.config = config
-
-    @abstractmethod
-    def get_layout(self) -> SpatialLayout:
-        """Produce spatial layout describing player↔pixel/token mapping.
-
-        Called once per image. Must NOT access the GPU.
-        """
-        ...
-
-    @abstractmethod
-    def generate_masks(
-        self,
-        coalitions_image: np.ndarray | None = None,
-        coalitions_text: np.ndarray | None = None,
-        device: torch.device | None = None,
-    ) -> PhysicalMask:
-        """Translate boolean coalition arrays into concrete physical masks.
-
-        Args:
-            coalitions_image: (N_img, n_players_image) bool.
-            coalitions_text: (N_txt, n_players_text) bool.
-
-        Returns:
-            PhysicalMask with image_binary_mask and/or text_attention_mask.
-
-        Note:
-            MUST NOT access the GPU inside this method
-            ("CPU Planning, GPU Execution").
-        """
-        ...
-
-    # ── Shared text-mask helper for VLM segmenters ──────────────────────
-
-    def _build_text_attention_mask(
-        self,
-        coalitions: np.ndarray,
-        device: torch.device | None = None,
-    ) -> torch.Tensor:
-        """Convert token-level coalitions to an attention mask matching the
-        model's expected total length.
-
-        Args:
-            coalitions: (N, n_players_text) bool.
-
-        Returns:
-            (N, text_total_length) int — 1=attend, 0=ignore.
-        """
-        cfg = self.config
-        coalition_t = torch.as_tensor(coalitions, dtype=torch.bool, device=device)
-        n_coalitions = coalition_t.shape[0]
-
-        if cfg.model_type in ("siglip", "siglip2"):
-            pad_len = cfg.text_total_length - cfg.n_players_text
-            return torch.cat(
-                (
-                    coalition_t,
-                    torch.ones(n_coalitions, pad_len, device=coalition_t.device),
-                ),
-                dim=1,
-            ).int()
-        if cfg.model_type == "clip":
-            return torch.cat(
-                (
-                    torch.ones(n_coalitions, 1, device=coalition_t.device),
-                    coalition_t,
-                    torch.ones(n_coalitions, 1, device=coalition_t.device),
-                ),
-                dim=1,
-            ).int()
-        raise ValueError(f"Unsupported model_type: {cfg.model_type}")
-
-
-class Masker(ABC):
-    """Abstract base class for feature occlusion strategies.
-
-    A Masker applies a PhysicalMask to model inputs (ProcessorOutput)
-    and returns modified inputs ready for model.forward().
-    """
-
-    def __init__(self, config: MaskerConfig | None = None):
-        self.config = config or MaskerConfig()
-
-    @abstractmethod
-    def apply(
-        self,
-        processor_output: ProcessorOutput,
-        physical_mask: PhysicalMask,
-    ) -> ProcessorOutput:
-        """Apply the physical mask to processor outputs.
-
-        Args:
-            processor_output: Preprocessed model inputs.
-            physical_mask: Concrete pixel/token masks to apply.
-
-        Returns:
-            Modified ProcessorOutput with occlusion applied.
-
-        Note:
-            Implementations MUST clone inputs before mutation.
-        """
-        ...
