+++ src/shapiq/imputer/vision/maskers/vision_mean.py
@@ -6,7 +6,8 @@ attention_mask.
 
 from __future__ import annotations
 
-from ..base import Masker, PhysicalMask, ProcessorOutput
+from ..base import PhysicalMask, ProcessorOutput
+from ..maskers.base import Masker
 from . import register_masker
