+++ src/shapiq/imputer/vision/segmenters/patch.py
@@ -9,7 +9,8 @@ from __future__ import annotations
 import numpy as np
 import torch
 
-from ..base import PhysicalMask, Segmenter, SegmenterConfig, SpatialLayout
+from ..base import PhysicalMask, SpatialLayout
+from ..segmenters.base import Segmenter, SegmenterConfig
 from . import register_segmenter
